+++ mesonbuild/compilers/__init__.py
@@ -35,6 +35,22 @@ __all__ = [
     'lang_suffixes',
     'sort_clink',
 
+    'compiler_from_language',
+    'detect_compiler_for',
+    'detect_static_linker',
+    'detect_c_compiler',
+    'detect_cpp_compiler',
+    'detect_cuda_compiler',
+    'detect_fortran_compiler',
+    'detect_objc_compiler',
+    'detect_objcpp_compiler',
+    'detect_java_compiler',
+    'detect_cs_compiler',
+    'detect_vala_compiler',
+    'detect_rust_compiler',
+    'detect_d_compiler',
+    'detect_swift_compiler',
+
     'AppleClangCCompiler',
     'AppleClangCPPCompiler',
     'AppleClangObjCCompiler',
@@ -132,6 +148,23 @@ from .compilers import (
     lang_suffixes,
     LANGUAGES_USING_LDFLAGS,
     sort_clink,
+)
+from .detect import (
+    compiler_from_language,
+    detect_compiler_for,
+    detect_static_linker,
+    detect_c_compiler,
+    detect_cpp_compiler,
+    detect_cuda_compiler,
+    detect_objc_compiler,
+    detect_objcpp_compiler,
+    detect_fortran_compiler,
+    detect_java_compiler,
+    detect_cs_compiler,
+    detect_vala_compiler,
+    detect_rust_compiler,
+    detect_d_compiler,
+    detect_swift_compiler,
 )
 from .c import (
     CCompiler,
